Move helper classes to new folder
