+++ helpers/ImagePreparer.ts
